Implemented specialized serialization of errors
This change makes use of the new `slog` feature which allows
specializing error types. When an error type is encountered, it can be
inspected to check if it contains an errno. If it does, the standard
`ERRNO=` field will be added. This however requires a newer version of
Rust (1.30+), so it's an opt-in feature.

Further, logging of error sources separately is supported. This is also
behind a feature flag as some users might find it redundant. Making it
runtime configurable would be nicer, but such would be a breaking change
because the `JournaldDrain` struct was not future-proofed. It should be
quite easy to make a new breaking release with this option though. This
feature calls `cause()` so it does not require a new version of Rust. To
aid log analyzers, it also logs the depth of the error chain.

Finally each error is logged as in the case of the default
implementation. The features described above do not interfere with this.
It could be also configurable but I'm not sure how to best achieve it.
The displaying code was copy-pasted from the `slog` crate because it's
not public. It's a good question whether it should be made public, or
even moved to a different crate.